# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a customized FizzBuzz server.",
        "title": "FizzBuzz API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/ping": {
            "get": {
                "description": "Returns pong",
                "produces": [
                    "application/json"
                ],
                "summary": "Ping",
                "operationId": "ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.StatusOKBuilder"
                        }
                    }
                }
            }
        },
        "/play": {
            "post": {
                "description": "Returns a customized fizz-buzz list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Play custom FizzBuzz",
                "operationId": "play-fizzbuzz",
                "parameters": [
                    {
                        "description": "Number1",
                        "name": "number1",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Number2",
                        "name": "number2",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Replace String1",
                        "name": "replace1",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Replace String2",
                        "name": "replace2",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Limit",
                        "name": "limit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PlayFizzBuzzResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.PlayFizzBuzzResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        },
        "utils.StatusOKBuilder": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}